requirements.txt
fastapi
uvicorn
black
sqlalchemy
passlib
bcrypt
python-jose
cryptography
python-multipart 
pymongo
motor==2.3.0
email-validator==1.1.2
uuid
random-username
python-dotenv